+++ docs/conf.py
@@ -1,12 +1,15 @@
 # Configuration file for the Sphinx documentation builder.
 import os, sys
 sys.path.insert(0, os.path.abspath('../src'))
-autodoc_mock_imports = ["torch", "torchvision"]
+
+# Mock imports so Sphinx doesn't crash if these aren't installed on the doc server
+autodoc_mock_imports = ["torch", "torchvision", "triton"] 
+
 exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store', '**/_build']
 project   = "SAGA"
 copyright = "2026, Siju K.S. et al."
 author    = "Siju K.S., Vipin Venugopal, Mithun Kumar Kar, Jayakrishnan Anandakrishnan"
-release   = "0.1.4"
+release   = "0.2.0"
 
 extensions = [
     "sphinx.ext.autodoc",